requirements.txt
# Core
sqlalchemy>=2.0
python-dotenv>=1.0

# API
fastapi>=0.100
uvicorn[standard]>=0.22
requests>=2.28

# Azure Key Vault (for scripts/keyvault_loader.py)
azure-identity>=1.15
azure-keyvault-secrets>=4.7

# Database drivers (install as needed)
# psycopg2-binary   # PostgreSQL
# pyodbc            # MSSQL (requires ODBC Driver)
# oracledb          # Oracle
